+++ @pew/pew-update-marca.php
@@ -11,10 +11,15 @@
             $invalid_fileds[$i] = $post_name;
         }
     }
+
     if($gravar){
         echo "<h3 align=center>Atualizando marca...</h3>";
+        
         require_once "pew-system-config.php";
+        require_once "@classe-system-functions.php";
+        
         $tabela_marcas = $pew_custom_db->tabela_marcas;
+        
         $idMarca = (int)$_POST["id_marca"];
         $titulo = addslashes(trim($_POST["titulo"]));
         $descricao = addslashes(trim($_POST["descricao"]));
@@ -22,21 +27,7 @@
         $data = date("Y-m-d h:i:s");
         $dirImagens = "../imagens/marcas/";
 
-        if(!function_exists("stringFormat")){
-            function stringFormat($string){
-                $string = str_replace("Ç", "c", $string);
-                $string = str_replace("ç", "c", $string);
-                $string = preg_replace(array("/(á|à|ã|â|ä)/","/(Á|À|Ã|Â|Ä)/","/(é|è|ê|ë)/","/(É|È|Ê|Ë)/","/(í|ì|î|ï)/","/(Í|Ì|Î|Ï)/","/(ó|ò|õ|ô|ö)/","/(Ó|Ò|Õ|Ô|Ö)/","/(ú|ù|û|ü)/","/(Ú|Ù|Û|Ü)/","/(ñ)/","/(Ñ)/"),explode(" ","a A e E i I o O u U n N"), $string);
-                $string = strtolower($string);
-                $string = str_replace("/", "-", $string);
-                $string = str_replace("|", "-", $string);
-                $string = str_replace(" ", "-", $string);
-                $string = str_replace(",", "", $string);
-                return $string;
-            }
-        }
-
-        $refMarca = stringFormat($titulo);
+        $refMarca = $pew_functions->url_format($titulo);
 
         $queryImagem = mysqli_query($conexao, "select imagem from $tabela_marcas where id = '$idMarca'");
         $infoImagem = mysqli_fetch_array($queryImagem);
@@ -48,7 +39,8 @@
             $ext = pathinfo($_FILES["imagem"]["name"], PATHINFO_EXTENSION);
             $nomeImgMarca = $nomeImgMarca.".".$ext;
             move_uploaded_file($_FILES["imagem"]["tmp_name"], $dirImagens.$nomeImgMarca);
-            if(file_exists($dirImagens.$imagemAtual)){
+            
+            if(file_exists($dirImagens.$imagemAtual) && $imagemAtual != ""){
                 unlink($dirImagens.$imagemAtual);
             }
         }else{
@@ -56,6 +48,7 @@
         }
 
         mysqli_query($conexao, "update $tabela_marcas set marca = '$titulo', descricao = '$descricao', ref = '$refMarca', imagem = '$nomeImgMarca', data_controle = '$data', status = '$status' where id = '$idMarca'");
+        
         echo "<script>window.location.href='pew-marcas.php?focus=$titulo&msg=A marca foi atualizada com sucesso!&msgType=success'</script>";
     }else{
         echo "<script>window.location.href='pew-marcas.php?msg=Ocorreu um erro ao atualizar a marca'</script>";
